requirements.txt
python-dotenv
google-adk
vertexai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib